Prod Push, 12-May

+++ public/admin/assets/js/ssb-site-builder/ssb-controls/ssb-edit-control/ssb-edit-control.controller.js
@@ -242,22 +242,52 @@ function ssbSiteBuilderEditControlController($scope, $rootScope, $interval, $att
     }
 
     function removeSectionFromPage(index) {
-        SweetAlert.swal({
-            title: "Are you sure?",
-            text: "Do you want to delete this section?",
-            type: "warning",
-            showCancelButton: true,
-            confirmButtonColor: "#DD6B55",
-            confirmButtonText: "Yes, delete it!",
-            cancelButtonText: "No, do not delete it!",
-            closeOnConfirm: true,
-            closeOnCancel: true
-        },
-        function (isConfirm) {
+        if(vm.state.page.sections[index].global){
+            SweetAlert.swal({
+                title: "Are you sure?",
+                text: "You are removing a global section. Changes made to global sections on this page will be reflected on all other pages. Consider removing from this page only.",
+                type: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "Remove from all pages",
+                cancelButtonText: "Hide on this page",
+                showNoActionButton: true,
+                noActionButtonText: 'Cancel',
+                closeOnConfirm: true,
+                closeOnCancel: true
+            },
+            function (isConfirm) {
+                //Remove from all pages
+                if (isConfirm) {
+                    vm.state.page.sections[index].global = false;
+                    vm.state.page.sections[index].visibility = false;
+                    vm.uiState.toggleSection(vm.state.page.sections[index]);
+                }
+                //Hide on this page
+                else if(angular.isDefined(isConfirm) && isConfirm === false){
+                    vm.state.page.sections[index].visibility = false;
+                    vm.uiState.toggleSection(vm.state.page.sections[index]);
+                }
+            });
+        }
+        else{
+            SweetAlert.swal({
+                title: "Are you sure?",
+                text: "Do you want to delete this section?",
+                type: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "Yes, delete it!",
+                cancelButtonText: "No, do not delete it!",
+                closeOnConfirm: true,
+                closeOnCancel: true
+            },
+            function (isConfirm) {
             if (isConfirm) {
                 SimpleSiteBuilderService.removeSectionFromPage(index)
             }
-        });
+            });
+        }
     }
 
     function scrollToActiveSection() {
